+++ packages/server/core/src/base/middlewares/customServer/index.ts
@@ -1,18 +1,25 @@
-import { ServerRoute } from '@modern-js/types';
+import {
+  ServerRoute,
+  UnstableMiddlewareContext,
+  UnstableMiddleware,
+} from '@modern-js/types';
 import { time } from '@modern-js/runtime-utils/time';
 import { ServerBase } from '../../serverBase';
 import { ServerHookRunner } from '../../../core/plugin';
-import { Middleware, ServerEnv } from '../../../core/server';
+import { Context, Middleware, ServerEnv } from '../../../core/server';
 import { transformResponse } from '../../utils';
 import { ServerReportTimings } from '../../constants';
 import type { ServerNodeEnv } from '../../adapters/node/hono';
+import { getLoaderCtx } from './loader';
 import {
   getAfterMatchCtx,
   getAfterRenderCtx,
   createCustomMiddlewaresCtx,
   createAfterStreamingRenderContext,
 } from './context';
-import { createBaseHookContext } from './base';
+import { ResArgs, createBaseHookContext } from './base';
+
+export { getLoaderCtx } from './loader';
 
 // eslint-disable-next-line @typescript-eslint/no-empty-function
 const noop = () => {};
@@ -48,7 +55,7 @@ export class CustomServer {
           middleware: webExtension,
         },
       },
-      { onLast: () => null },
+      { onLast: () => [] },
     );
   }
 
@@ -156,19 +163,37 @@ export class CustomServer {
     };
   }
 
-  getServerMiddleware(): Middleware<ServerNodeEnv & ServerEnv> {
+  async getServerMiddleware(): Promise<
+    | Middleware<ServerNodeEnv & ServerEnv>
+    | Array<Middleware<ServerNodeEnv & ServerEnv>>
+    | undefined
+  > {
+    const serverMiddleware = await this.serverMiddlewarePromise;
+
+    if (!serverMiddleware) {
+      return;
+    }
+
+    if (Array.isArray(serverMiddleware)) {
+      // eslint-disable-next-line consistent-return
+      return getUnstableMiddlewares(serverMiddleware);
+    }
+
     // eslint-disable-next-line consistent-return
     return async (c, next) => {
-      const serverMiddleware = await this.serverMiddlewarePromise;
-      if (!serverMiddleware) {
-        return next();
-      }
-
       const reporter = c.get('reporter');
 
       const locals: Record<string, any> = {};
 
-      const customMiddlewareCtx = createCustomMiddlewaresCtx(c, locals);
+      const resArgs: ResArgs = {
+        headers: new Headers(),
+      };
+
+      const customMiddlewareCtx = createCustomMiddlewaresCtx(
+        c,
+        locals,
+        resArgs,
+      );
 
       const getCost = time();
       await serverMiddleware(customMiddlewareCtx);
@@ -177,6 +202,13 @@ export class CustomServer {
         reporter?.reportTiming(ServerReportTimings.SERVER_MIDDLEWARE, cost);
 
       c.set('locals', locals);
+
+      if (isRedirect(resArgs.headers, resArgs.status)) {
+        return c.redirect(
+          resArgs.headers.get('Location') || '',
+          resArgs.status || 302,
+        );
+      }
 
       if (c.env?.node.res.headersSent) {
         return undefined;
@@ -187,4 +219,58 @@ export class CustomServer {
       }
     };
   }
+}
+
+function isRedirect(headers: Headers, code?: number) {
+  return [301, 302, 307, 308].includes(code || 0) || headers.get('Location');
+}
+
+function getUnstableMiddlewares(
+  serverMiddleware: UnstableMiddleware[],
+): Array<Middleware<ServerNodeEnv & ServerEnv>> {
+  return serverMiddleware.map(middleware => {
+    return async (c, next) => {
+      const context = createMiddlewareContextFromHono(c);
+
+      return middleware(context, next);
+    };
+  });
+}
+
+function createMiddlewareContextFromHono(
+  c: Context,
+): UnstableMiddlewareContext {
+  const loaderContext = getLoaderCtx(c);
+
+  return {
+    get request() {
+      return c.req.raw;
+    },
+
+    get response() {
+      return c.res;
+    },
+
+    set response(newRes) {
+      c.res = newRes;
+    },
+
+    get(key) {
+      return loaderContext.get(key as string);
+    },
+
+    set(key, value) {
+      return loaderContext.set(key as string, value);
+    },
+
+    status: c.status.bind(c),
+
+    header: c.header.bind(c),
+
+    body: c.body.bind(c),
+
+    html: c.html.bind(c),
+
+    redirect: c.redirect.bind(c),
+  };
 }