+++ packages/module/plugin-module-import/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @modern-js/plugin-module-import
 
+## 2.49.4
+
+### Patch Changes
+
+- @modern-js/module-tools@2.49.4
+
 ## 2.49.3
 
 ### Patch Changes