+++ packages/solutions/monorepo-tools/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @modern-js/monorepo-tools
 
+## 2.49.4
+
+### Patch Changes
+
+- @modern-js/core@2.49.4
+- @modern-js/plugin-changeset@2.49.4
+- @modern-js/plugin-i18n@2.49.4
+- @modern-js/plugin-lint@2.49.4
+- @modern-js/plugin@2.49.4
+- @modern-js/utils@2.49.4
+
 ## 2.49.3
 
 ### Patch Changes